requirements.txt
flask==3.0.0
python-dotenv==1.0.0
google-genai==0.3.0
ffmpeg-python==0.2.0
werkzeug==3.0.1
demucs>=4.0.0
torch>=1.9.0
torchaudio>=0.9.0
librosa>=0.8.0
soundfile>=0.10.0
numpy>=1.19.0
scipy>=1.7.0
google-cloud-texttospeech>=2.16.0
